pyproject.toml
[tool.poetry]
name = "weather-disaster-adk"
version = "1.0.0"
description = "Weather Disaster Management System using Google ADK"
authors = ["Adithya"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
google-adk = "^0.1.0"
google-genai = "^0.2.0"
google-cloud-aiplatform = "^1.38.0"
google-generativeai = "^0.3.0"
pydantic = "^2.5.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
structlog = "^23.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
black = "^23.12.0"
isort = "^5.13.0"
flake8 = "^6.1.0"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=term-missing"
